+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.5.77] - 2020-01-12
+### Added
+- Ability to convert DOCX to PDF using CloudConvert.
+- The `cloudconvert secret` Configuration directive.
+
 ## [0.5.76] - 2020-01-08
 ### Fixed
 - Playground variables sidebar did not always switch to the most